fix: allow vaulted passphrases
Using a vault for the ca_passphrase results in the var being an object. The current assertion that ca_passphrase is a string will fail then.

The ansible team approved a pull request which added the capability to test for vaulted values back in 2020. I think it is this one: ansible/ansible#67823

Adding the or conjunction allows using plain text passphrases as well as vaulted ones.